Delete CustomEvents from poiner input and a few other APIs
This CL deleted custom events infrastructure as well as other related APIs: scroll orientation locking, pointer input filter initialisation with custom events dispatcher, etc.

Relnote: CustomEvens support from PointerInputModifier has been removed
Test: pass
Bug: 175294473
