+++ compose/ui/ui/src/androidAndroidTest/kotlin/androidx/compose/ui/input/pointer/HitPathTrackerTest.kt
@@ -197,28 +197,6 @@ class HitPathTrackerTest {
     }
 
     @Test
-    fun addHitPath_1NodeAdded_initHandlerCalledWithValidCustomMessageDispatcher() {
-        val pif = PointerInputFilterMock()
-
-        hitPathTracker.addHitPath(PointerId(3), listOf(pif))
-
-        assertThat(pif.log.getOnInitLog()).hasSize(1)
-    }
-
-    @Test
-    fun addHitPath_3NodesAdded_allInitHandlersCalledWithValidCustomMessageDispatcher() {
-        val pifParent = PointerInputFilterMock()
-        val pifMiddle = PointerInputFilterMock()
-        val pifChild = PointerInputFilterMock()
-
-        hitPathTracker.addHitPath(PointerId(3), listOf(pifParent, pifMiddle, pifChild))
-
-        assertThat(pifParent.log.getOnInitLog()).hasSize(1)
-        assertThat(pifMiddle.log.getOnInitLog()).hasSize(1)
-        assertThat(pifChild.log.getOnInitLog()).hasSize(1)
-    }
-
-    @Test
     fun dispatchChanges_noNodes_doesNotCrash() {
         hitPathTracker.dispatchChanges(internalPointerEventOf(down(0)))
     }
@@ -2950,231 +2928,8 @@ class HitPathTrackerTest {
         assertThat(areEqual(hitPathTracker.root, NodeParent())).isTrue()
     }
 
-    @Test
-    fun dispatchCustomEvent_1NodeItDispatches_nothingReceivesDispatch() {
-
-        // Arrange
-
-        lateinit var dispatcher: CustomEventDispatcher
-
-        val pif = PointerInputFilterMock(
-            initHandler = { dispatcher = it }
-        )
-
-        hitPathTracker.addHitPath(PointerId(3), listOf(pif))
-
-        val event = TestCustomEvent("test")
-
-        // Act
-
-        dispatcher.dispatchCustomEvent(event)
-
-        // Assert
-
-        assertThat(pif.log.getOnCustomEventLog()).hasSize(0)
-    }
-
-    @Test
-    fun dispatchCustomEvent_1Path3NodesParentDispatches_dispatchCorrect() {
-        dispatchCustomEvent_1Path3Nodes_dispatchCorrect(DispatchingPif.Parent)
-    }
-
-    @Test
-    fun dispatchCustomEvent_1Path3NodesMiddleDispatches_dispatchCorrect() {
-        dispatchCustomEvent_1Path3Nodes_dispatchCorrect(DispatchingPif.Middle)
-    }
-
-    @Test
-    fun dispatchCustomEvent_1Path3NodesChildDispatches_dispatchCorrect() {
-        dispatchCustomEvent_1Path3Nodes_dispatchCorrect(DispatchingPif.Child)
-    }
-
     private enum class DispatchingPif {
         Parent, Middle, Child
-    }
-
-    private fun dispatchCustomEvent_1Path3Nodes_dispatchCorrect(
-        dispatchingPif: DispatchingPif
-    ) {
-        // Arrange
-
-        lateinit var dispatcher: CustomEventDispatcher
-        lateinit var parentPif: PointerInputFilter
-        lateinit var middlePif: PointerInputFilter
-        lateinit var childPif: PointerInputFilter
-
-        lateinit var seniorPif: PointerInputFilter
-        lateinit var juniorPif: PointerInputFilter
-        val dispatcherInitHandler: (CustomEventDispatcher) -> Unit = { dispatcher = it }
-
-        val log = mutableListOf<LogEntry>()
-
-        when (dispatchingPif) {
-            DispatchingPif.Parent -> {
-                parentPif = PointerInputFilterMock(
-                    log,
-                    initHandler = dispatcherInitHandler
-                )
-                middlePif = PointerInputFilterMock(log)
-                seniorPif = middlePif
-                childPif = PointerInputFilterMock(log)
-                juniorPif = childPif
-            }
-            DispatchingPif.Middle -> {
-                parentPif = PointerInputFilterMock(log)
-                seniorPif = parentPif
-                middlePif = PointerInputFilterMock(
-                    log,
-                    initHandler = dispatcherInitHandler
-                )
-                childPif = PointerInputFilterMock(log)
-                juniorPif = childPif
-            }
-            DispatchingPif.Child -> {
-                parentPif = PointerInputFilterMock(log)
-                seniorPif = parentPif
-                middlePif = PointerInputFilterMock(log)
-                juniorPif = middlePif
-                childPif = PointerInputFilterMock(
-                    log,
-                    initHandler = dispatcherInitHandler
-                )
-            }
-        }
-
-        hitPathTracker.addHitPath(PointerId(3), listOf(parentPif, middlePif, childPif))
-
-        val event = TestCustomEvent("test")
-
-        // Act
-
-        dispatcher.dispatchCustomEvent(event)
-
-        // Assert
-
-        val log1 = log.getOnCustomEventLog()
-
-        assertThat(log1).hasSize(6)
-
-        assertThat(log1[0].pointerInputFilter).isEqualTo(seniorPif)
-        assertThat(log1[0].pass).isEqualTo(PointerEventPass.Initial)
-        assertThat(log1[1].pointerInputFilter).isEqualTo(juniorPif)
-        assertThat(log1[1].pass).isEqualTo(PointerEventPass.Initial)
-        assertThat(log1[2].pointerInputFilter).isEqualTo(juniorPif)
-        assertThat(log1[2].pass).isEqualTo(PointerEventPass.Main)
-        assertThat(log1[3].pointerInputFilter).isEqualTo(seniorPif)
-        assertThat(log1[3].pass).isEqualTo(PointerEventPass.Main)
-        assertThat(log1[4].pointerInputFilter).isEqualTo(seniorPif)
-        assertThat(log1[4].pass).isEqualTo(PointerEventPass.Final)
-        assertThat(log1[5].pointerInputFilter).isEqualTo(juniorPif)
-        assertThat(log1[5].pass).isEqualTo(PointerEventPass.Final)
-    }
-
-    @Test
-    fun dispatchCustomEvent_1Parent2ChildrenParentDispatches_dispatchCorrect() {
-
-        lateinit var dispatcher: CustomEventDispatcher
-
-        val log = mutableListOf<LogEntry>()
-
-        val parentPin = PointerInputFilterMock(
-            log,
-            initHandler = { dispatcher = it }
-        )
-        val childPin1 = PointerInputFilterMock(log)
-        val childPin2 = PointerInputFilterMock(log)
-
-        hitPathTracker.addHitPath(PointerId(3), listOf(parentPin, childPin1))
-        hitPathTracker.addHitPath(PointerId(4), listOf(parentPin, childPin2))
-
-        val event = TestCustomEvent("test")
-
-        // Act
-
-        dispatcher.dispatchCustomEvent(event)
-
-        // Assert
-
-        val log1 = log.getOnCustomEventLog()
-
-        assertThat(log1).hasSize(6)
-
-        assertThat(log1[0].pointerInputFilter).isEqualTo(childPin1)
-        assertThat(log1[0].pass).isEqualTo(PointerEventPass.Initial)
-        assertThat(log1[1].pointerInputFilter).isEqualTo(childPin1)
-        assertThat(log1[1].pass).isEqualTo(PointerEventPass.Main)
-        assertThat(log1[2].pointerInputFilter).isEqualTo(childPin2)
-        assertThat(log1[2].pass).isEqualTo(PointerEventPass.Initial)
-        assertThat(log1[3].pointerInputFilter).isEqualTo(childPin2)
-        assertThat(log1[3].pass).isEqualTo(PointerEventPass.Main)
-
-        assertThat(log1[4].pointerInputFilter).isEqualTo(childPin1)
-        assertThat(log1[4].pass).isEqualTo(PointerEventPass.Final)
-        assertThat(log1[5].pointerInputFilter).isEqualTo(childPin2)
-        assertThat(log1[5].pass).isEqualTo(PointerEventPass.Final)
-    }
-
-    @Test
-    fun dispatchCustomEvent_1Parent2ChildrenChild1Dispatches_dispatchCorrect() {
-        dispatchCustomEvent_1Parent2ChildrenChildDispatches_dispatchCorrect(
-            true
-        )
-    }
-
-    @Test
-    fun dispatchCustomEvent_1Parent2ChildrenChild2Dispatches_dispatchCorrect() {
-        dispatchCustomEvent_1Parent2ChildrenChildDispatches_dispatchCorrect(
-            false
-        )
-    }
-
-    private fun dispatchCustomEvent_1Parent2ChildrenChildDispatches_dispatchCorrect(
-        firstChildDispatches: Boolean
-    ) {
-        // Arrange
-
-        val log = mutableListOf<LogEntry>()
-
-        val parentPif = PointerInputFilterMock(log)
-        lateinit var childPif1: PointerInputFilter
-        lateinit var childPif2: PointerInputFilter
-
-        lateinit var dispatcher: CustomEventDispatcher
-        val initHandler: (CustomEventDispatcher) -> Unit = { dispatcher = it }
-
-        if (firstChildDispatches) {
-            childPif1 = PointerInputFilterMock(
-                log,
-                initHandler = initHandler
-            )
-            childPif2 = PointerInputFilterMock(log)
-        } else {
-            childPif1 = PointerInputFilterMock(log)
-            childPif2 = PointerInputFilterMock(
-                log,
-                initHandler = initHandler
-            )
-        }
-
-        hitPathTracker.addHitPath(PointerId(3), listOf(parentPif, childPif1))
-        hitPathTracker.addHitPath(PointerId(4), listOf(parentPif, childPif2))
-
-        val event = TestCustomEvent("test")
-
-        // Act
-
-        dispatcher.dispatchCustomEvent(event)
-
-        val log1 = log.getOnCustomEventLog()
-
-        assertThat(log1).hasSize(3)
-
-        assertThat(log1[0].pointerInputFilter).isEqualTo(parentPif)
-        assertThat(log1[0].pass).isEqualTo(PointerEventPass.Initial)
-        assertThat(log1[1].pointerInputFilter).isEqualTo(parentPif)
-        assertThat(log1[1].pass).isEqualTo(PointerEventPass.Main)
-        assertThat(log1[2].pointerInputFilter).isEqualTo(parentPif)
-        assertThat(log1[2].pass).isEqualTo(PointerEventPass.Final)
     }
 
     // Tests related to reporting whether or not a pointer input filter was dispatched to.
@@ -3206,246 +2961,6 @@ class HitPathTrackerTest {
     }
 
     // Tests related to retaining and releasing hit paths.
-
-    /**
-     * Verifies that if a hit path is added and retained, when it is removed, it is not actually
-     * removed.
-     */
-    @Test
-    fun removeHitPath_idRetained_nodeIsRetained() {
-        lateinit var dispatcher: CustomEventDispatcher
-        val pif = PointerInputFilterMock(
-            initHandler = { dispatcher = it }
-        )
-        hitPathTracker.addHitPath(PointerId(13), listOf(pif))
-        dispatcher.retainHitPaths(setOf(PointerId(13)))
-
-        hitPathTracker.removeHitPath(PointerId(13))
-
-        val expectedRoot = NodeParent().apply {
-            children.add(
-                Node(pif).apply {
-                    pointerIds.add(PointerId(13))
-                }
-            )
-        }
-
-        assertThat(areEqual(hitPathTracker.root, expectedRoot)).isTrue()
-    }
-
-    /**
-     * Verifies that if there are 2 branching hit paths, 1 is retained, and both are removed, the
-     * 1 that was retained remains.
-     */
-    @Test
-    fun removeHitPath_2Branches1RetainedBothRemoved_retainedBranchRemains() {
-        lateinit var dispatcher: CustomEventDispatcher
-        val parentPif = PointerInputFilterMock()
-        val childPif1 = PointerInputFilterMock()
-        val childPif2 = PointerInputFilterMock(
-            initHandler = { dispatcher = it }
-        )
-
-        hitPathTracker.addHitPath(PointerId(1), listOf(parentPif, childPif1))
-        hitPathTracker.addHitPath(PointerId(2), listOf(parentPif, childPif2))
-        dispatcher.retainHitPaths(setOf(PointerId(1)))
-
-        hitPathTracker.removeHitPath(PointerId(1))
-        hitPathTracker.removeHitPath(PointerId(2))
-
-        val expectedRoot = NodeParent().apply {
-            children.add(
-                Node(parentPif).apply {
-                    pointerIds.add(PointerId(1))
-                    children.add(
-                        Node(childPif1).apply {
-                            pointerIds.add(PointerId(1))
-                        }
-                    )
-                }
-            )
-        }
-
-        assertThat(areEqual(hitPathTracker.root, expectedRoot)).isTrue()
-    }
-
-    /**
-     * Verifies that if there are 2 branching hit paths, both are retained, and both are removed,
-     * that both remain.
-     */
-    @Test
-    fun removeHitPath_2Branches2RetainedBothRemoved_bothBranchesRemain() {
-        lateinit var dispatcher: CustomEventDispatcher
-        val parentPif = PointerInputFilterMock(
-            initHandler = { dispatcher = it }
-        )
-        val childPif1 = PointerInputFilterMock()
-        val childPif2 = PointerInputFilterMock()
-
-        hitPathTracker.addHitPath(PointerId(1), listOf(parentPif, childPif1))
-        hitPathTracker.addHitPath(PointerId(2), listOf(parentPif, childPif2))
-        dispatcher.retainHitPaths(setOf(PointerId(1), PointerId(2)))
-
-        hitPathTracker.removeHitPath(PointerId(1))
-        hitPathTracker.removeHitPath(PointerId(2))
-
-        val expectedRoot = NodeParent().apply {
-            children.add(
-                Node(parentPif).apply {
-                    pointerIds.addAll(listOf(PointerId(1), PointerId(2)))
-                    children.add(
-                        Node(childPif1).apply {
-                            pointerIds.add(PointerId(1))
-                        }
-                    )
-                    children.add(
-                        Node(childPif2).apply {
-                            pointerIds.add(PointerId(2))
-                        }
-                    )
-                }
-            )
-        }
-
-        assertThat(areEqual(hitPathTracker.root, expectedRoot)).isTrue()
-    }
-
-    /**
-     * Verifies that if a hit path is retained, then removed, then released, it is removed.
-     */
-    @Test
-    fun releasePointerId_idRetainedAndPathRemoved_nodeIsRemoved() {
-        lateinit var dispatcher: CustomEventDispatcher
-        val pif = PointerInputFilterMock(
-            initHandler = { dispatcher = it }
-        )
-        hitPathTracker.addHitPath(PointerId(13), listOf(pif))
-        dispatcher.retainHitPaths(setOf(PointerId(13)))
-        hitPathTracker.removeHitPath(PointerId(13))
-
-        dispatcher.releaseHitPaths(setOf(PointerId(13)))
-
-        assertThat(areEqual(hitPathTracker.root, NodeParent())).isTrue()
-    }
-
-    /**
-     * Verifies that if there are 2 hit paths, both are retained, both are removed, and
-     * then 1 is released, the other retained branch remains.
-     */
-    @Test
-    fun releasePointerId_2Branches2RetainedBothRemoved1Released_correctBranchRemains() {
-        lateinit var dispatcher: CustomEventDispatcher
-        val parentPif = PointerInputFilterMock(
-            initHandler = { dispatcher = it }
-        )
-        val childPif1 = PointerInputFilterMock()
-        val childPif2 = PointerInputFilterMock()
-
-        hitPathTracker.addHitPath(PointerId(1), listOf(parentPif, childPif1))
-        hitPathTracker.addHitPath(PointerId(2), listOf(parentPif, childPif2))
-        dispatcher.retainHitPaths(setOf(PointerId(1), PointerId(2)))
-
-        hitPathTracker.removeHitPath(PointerId(1))
-        hitPathTracker.removeHitPath(PointerId(2))
-
-        dispatcher.releaseHitPaths(setOf(PointerId(1)))
-
-        val expectedRoot = NodeParent().apply {
-            children.add(
-                Node(parentPif).apply {
-                    pointerIds.add(PointerId(2))
-                    children.add(
-                        Node(childPif2).apply {
-                            pointerIds.add(PointerId(2))
-                        }
-                    )
-                }
-            )
-        }
-
-        assertThat(areEqual(hitPathTracker.root, expectedRoot)).isTrue()
-    }
-
-    /**
-     * Verifies that if there are 2 hit paths, 1 is retained, they are both removed, and
-     * then the one that was retained is released, that no branches remain.
-     */
-    @Test
-    fun releasePointerId_2Branches1RetainedBothRemoved1Released_noBranchesRemain() {
-        lateinit var dispatcher: CustomEventDispatcher
-        val parentPif = PointerInputFilterMock()
-        val childPif1 = PointerInputFilterMock(
-            initHandler = { dispatcher = it }
-        )
-        val childPif2 = PointerInputFilterMock()
-
-        hitPathTracker.addHitPath(PointerId(1), listOf(parentPif, childPif1))
-        hitPathTracker.addHitPath(PointerId(2), listOf(parentPif, childPif2))
-        dispatcher.retainHitPaths(setOf(PointerId(1)))
-
-        hitPathTracker.removeHitPath(PointerId(1))
-        hitPathTracker.removeHitPath(PointerId(2))
-
-        dispatcher.releaseHitPaths(setOf(PointerId(1)))
-
-        assertThat(areEqual(hitPathTracker.root, NodeParent())).isTrue()
-    }
-
-    /**
-     * Verifies that if a hit path is retained and then released (without it ever actually being
-     * removed) then the hit path remains.
-     */
-    @Test
-    fun releasePointerId_idRetainedButPathNotRemoved_pathNotRemoved() {
-        lateinit var dispatcher: CustomEventDispatcher
-        val pif = PointerInputFilterMock(
-            initHandler = { dispatcher = it }
-        )
-        hitPathTracker.addHitPath(PointerId(13), listOf(pif))
-        dispatcher.retainHitPaths(setOf(PointerId(13)))
-
-        dispatcher.releaseHitPaths(setOf(PointerId(13)))
-
-        val expectedRoot = NodeParent().apply {
-            children.add(
-                Node(pif).apply {
-                    pointerIds.add(PointerId(13))
-                }
-            )
-        }
-
-        assertThat(areEqual(hitPathTracker.root, expectedRoot)).isTrue()
-    }
-
-    /**
-     * Verifies that if a hit path is retained, and then removed, and a dispatch of a
-     * custom event occurs, it will be dispatched to the retained path.
-     */
-    @Test
-    fun dispatchCustomEvent_idRetainedAndPathRemoved_customEventReachesNode() {
-        val log = mutableListOf<LogEntry>()
-        lateinit var dispatcher: CustomEventDispatcher
-        val pif = PointerInputFilterMock(
-            log,
-            initHandler = { dispatcher = it }
-        )
-        val pif2 = PointerInputFilterMock(log)
-        hitPathTracker.addHitPath(PointerId(13), listOf(pif, pif2))
-        dispatcher.retainHitPaths(setOf(PointerId(13)))
-        hitPathTracker.removeHitPath(PointerId(13))
-        val event = TestCustomEvent("87483")
-
-        dispatcher.dispatchCustomEvent(event)
-
-        val log1 = log.getOnCustomEventLog()
-        assertThat(log1).hasSize(3)
-        assertThat(log1[0].pointerInputFilter).isEqualTo(pif2)
-        assertThat(log1[0].pass).isEqualTo(PointerEventPass.Initial)
-        assertThat(log1[1].pointerInputFilter).isEqualTo(pif2)
-        assertThat(log1[1].pass).isEqualTo(PointerEventPass.Main)
-        assertThat(log1[2].pointerInputFilter).isEqualTo(pif2)
-        assertThat(log1[2].pass).isEqualTo(PointerEventPass.Final)
-    }
 
     @Test
     fun dispatchChanges_pifRemovesSelfDuringInitial_noPassesReceivedAfterwards() {
@@ -3682,197 +3197,6 @@ class HitPathTrackerTest {
         )
     }
 
-    @Test
-    fun dispatchCustomMessage_pifRemovesSelfDuringInitial_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovesSelfDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Initial
-        )
-    }
-
-    @Test
-    fun dispatchCustomMessage_pifRemovesSelfDuringMain_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovesSelfDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Main
-        )
-    }
-
-    @Test
-    fun dispatchCustomMessage_pifRemovesSelfDuringFinal_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovesSelfDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Final
-        )
-    }
-
-    private fun dispatchCustomMessage_pifRemovesSelfDuringDispatch_noPassesReceivedAfterwards(
-        removalPass: PointerEventPass
-    ) {
-
-        val log = mutableListOf<LogEntry>()
-
-        lateinit var dispatcher: CustomEventDispatcher
-
-        val layoutCoordinates = LayoutCoordinatesStub(true)
-
-        val dispatchingPif = PointerInputFilterMock(
-            log,
-            initHandler = { dispatcher = it }
-        )
-        val receivingPif = PointerInputFilterMock(
-            log,
-            onCustomEvent = { _, pointerEventPass ->
-                if (pointerEventPass == removalPass) {
-                    layoutCoordinates.isAttached = false
-                }
-            },
-            layoutCoordinates = layoutCoordinates
-        )
-
-        hitPathTracker.addHitPath(PointerId(13), listOf(dispatchingPif, receivingPif))
-
-        dispatcher.dispatchCustomEvent(object : CustomEvent {})
-
-        val log1 = log.getOnCustomEventLog().filter { it.pointerInputFilter == receivingPif }
-        val count = removalPass.ordinal + 1
-        assertThat(log1).hasSize(count)
-        PointerEventPass.values().forEachIndexed { index, pass ->
-            if (index < count) {
-                assertThat(log1[index].pass).isEqualTo(pass)
-            }
-        }
-    }
-
-    @Test
-    fun dispatchCustomMessage_pifRemovedByParentDuringInitial_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovedByParentDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Initial
-        )
-    }
-
-    @Test
-    fun dispatchCustomMessage_pifRemovedByParentDuringMain_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovedByParentDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Main
-        )
-    }
-
-    @Test
-    fun dispatchCustomMessage_pifRemovedByParentDuringFinal_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovedByParentDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Final
-        )
-    }
-
-    private fun dispatchCustomMessage_pifRemovedByParentDuringDispatch_noPassesReceivedAfterwards(
-        removalPass: PointerEventPass
-    ) {
-        lateinit var dispatcher: CustomEventDispatcher
-
-        val log = mutableListOf<LogEntry>()
-
-        val layoutCoordinates = LayoutCoordinatesStub(true)
-
-        val dispatchingPif = PointerInputFilterMock(
-            log,
-            initHandler = { dispatcher = it }
-        )
-        val parentPif = PointerInputFilterMock(
-            log,
-            onCustomEvent = { _, pointerEventPass ->
-                if (pointerEventPass == removalPass) {
-                    layoutCoordinates.isAttached = false
-                }
-            }
-        )
-        val childPif = PointerInputFilterMock(
-            log,
-            layoutCoordinates = layoutCoordinates
-        )
-
-        hitPathTracker.addHitPath(PointerId(13), listOf(dispatchingPif, parentPif, childPif))
-
-        dispatcher.dispatchCustomEvent(object : CustomEvent {})
-
-        val log1 = log.getOnCustomEventLog().filter { it.pointerInputFilter == childPif }
-        val count =
-            when (removalPass) {
-                PointerEventPass.Initial -> 0
-                PointerEventPass.Main -> 2
-                PointerEventPass.Final -> 2
-            }
-        assertThat(log1).hasSize(count)
-        PointerEventPass.values().forEachIndexed { index, pass ->
-            if (index < count) {
-                assertThat(log1[index].pass).isEqualTo(pass)
-            }
-        }
-    }
-
-    @Test
-    fun dispatchCustomMessage_pifRemovedByChildDuringInitial_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovedByChildDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Initial
-        )
-    }
-
-    @Test
-    fun dispatchCustomMessage_pifRemovedByChildDuringMain_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovedByChildDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Main
-        )
-    }
-
-    @Test
-    fun dispatchCustomMessage_pifRemovedByChildDuringFinal_noPassesReceivedAfterwards() {
-        dispatchCustomMessage_pifRemovedByChildDuringDispatch_noPassesReceivedAfterwards(
-            PointerEventPass.Final
-        )
-    }
-
-    private fun dispatchCustomMessage_pifRemovedByChildDuringDispatch_noPassesReceivedAfterwards(
-        removalPass: PointerEventPass
-    ) {
-        lateinit var dispatcher: CustomEventDispatcher
-
-        val log = mutableListOf<LogEntry>()
-
-        val layoutCoordinates = LayoutCoordinatesStub(true)
-
-        val dispatchingPif = PointerInputFilterMock(
-            log,
-            initHandler = { dispatcher = it }
-        )
-        val parentPif = PointerInputFilterMock(
-            log,
-            layoutCoordinates = layoutCoordinates
-        )
-        val childPif = PointerInputFilterMock(
-            log,
-            onCustomEvent = { _, pointerEventPass ->
-                if (pointerEventPass == removalPass) {
-                    layoutCoordinates.isAttached = false
-                }
-            }
-        )
-
-        hitPathTracker.addHitPath(PointerId(13), listOf(dispatchingPif, parentPif, childPif))
-
-        dispatcher.dispatchCustomEvent(object : CustomEvent {})
-
-        val log1 = log.getOnCustomEventLog().filter { it.pointerInputFilter == parentPif }
-        val count =
-            when (removalPass) {
-                PointerEventPass.Initial -> 1
-                PointerEventPass.Main -> 1
-                PointerEventPass.Final -> 3
-            }
-        assertThat(log1).hasSize(count)
-        PointerEventPass.values().forEachIndexed { index, pass ->
-            if (index < count) {
-                assertThat(log1[index].pass).isEqualTo(pass)
-            }
-        }
-    }
-
     private fun areEqual(actualNode: NodeParent, expectedNode: NodeParent): Boolean {
         var check = true
 
@@ -3916,8 +3240,6 @@ class HitPathTrackerTest {
         return check
     }
 }
-
-internal data class TestCustomEvent(val value: String) : CustomEvent
 
 class LayoutCoordinatesStub(
     override var isAttached: Boolean = true
