+++ compose/ui/ui/src/commonMain/kotlin/androidx/compose/ui/input/pointer/PointerEvent.kt
@@ -71,29 +71,6 @@ abstract class PointerInputFilter {
      */
     abstract fun onCancel()
 
-    /**
-     * Invoked right after this [PointerInputFilter] is hit by a pointer during hit testing.
-     *
-     * @param customEventDispatcher The [CustomEventDispatcher] that can be used to dispatch
-     * [CustomEvent] across the tree of hit [PointerInputFilter]s.
-     *
-     * @See CustomEventDispatcher
-     */
-    open fun onInit(customEventDispatcher: CustomEventDispatcher) {}
-
-    /**
-     * Invoked when a [CustomEvent] is dispatched by a [PointerInputFilter].
-     *
-     * Dispatch occurs over all passes of [PointerEventPass].
-     *
-     * @param customEvent The [CustomEvent] is the event being dispatched.
-     * @param pass The [PointerEventPass] in which this function is being called.
-     *
-     * @see CustomEvent
-     * @see PointerEventPass
-     */
-    open fun onCustomEvent(customEvent: CustomEvent, pass: PointerEventPass) {}
-
     internal var layoutCoordinates: LayoutCoordinates? = null
 
     /**
@@ -270,51 +247,6 @@ class ConsumedData(
  */
 enum class PointerEventPass {
     Initial, Main, Final
-}
-
-/**
- * The base type for all custom events.
- */
-interface CustomEvent
-
-/**
- * Defines the interface that is used to dispatch CustomEvents to pointer input nodes across the
- * compose tree.
- */
-interface CustomEventDispatcher {
-
-    /**
-     * Dispatches the [event] to all other pointer input nodes that share associated [PointerId]s
-     * with the pointer input node doing the dispatching.
-     *
-     * @param event The [CustomEvent] to dispatch.
-     */
-    // TODO(shepshapard): Come back and consider any issues with: This effectively allows
-    //  individual pointer input nodes to gain a reference back to the internal HitPathTracker.
-    //  But I think that is ok since pointer input nodes should  never be able to live for longer
-    //  than the HitPathTracker that would be responsible for tracking them.
-    fun dispatchCustomEvent(event: CustomEvent)
-
-    /**
-     * Arranges to retain the hit paths associated with the provided [pointerIds] such that if
-     * they are requested to be removed for any reason, they are retained.
-     *
-     * For example, this is useful when a pointer input filter wants to be able to send future
-     * custom messages to a another after the pointer has actually be released from the screen
-     * (such as in the case where a Double Tap gesture detector may want to delay a Single Tap
-     * gesture detector from firing but later may allow it to do so even after the pointer
-     * associated with the Single Tap Gesture detector no longer exists.
-     */
-    fun retainHitPaths(pointerIds: Set<PointerId>)
-
-    /**
-     * Arranges to release any hit paths associated with the provided [pointerIds] such that if
-     * they will be requested to be removed in the future, they will be removed upon request.
-     *
-     * If they were already requested to be removed while they were retained, they will be
-     * removed immediately upon release.
-     */
-    fun releaseHitPaths(pointerIds: Set<PointerId>)
 }
 
 /**
